COMP: Fix warnings bad conversion
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=15427
git-svn-id: http://svn.slicer.org/Slicer4/trunk@15427 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/Testing/vtkMRMLSceneTest2.cxx
@@ -440,7 +440,7 @@ int vtkMRMLSceneTest2(int argc, char * argv [] )
   // If additional nodes have been instanciated, let's update expectedNumberOfNode
   if (static_cast<int>(nodeAddedClassNames.size()) > expectedNumberOfNode)
     {
-    expectedNumberOfNode = nodeAddedClassNames.size();
+    expectedNumberOfNode = static_cast<int>(nodeAddedClassNames.size());
     }
 
   SCENE_CHECK_NUMBER_OF_EVENT("Problem with vtkMRMLScene::Connect()", "NumberOfNode",
